'pass the ball'
develop a framework to pass an object (a ball) between objects (players) standing in an object (the environment). players should be able to see the ball when it is in the environment, by accessing a list of present objects (including the players themselves). take possession of the ball, by adding a reference to it in their inventory property and tell the environment to remove it from the list of items present. while in possession of the ball, players can change its color and pick a new player to toss it to.

right now i'm just getting hung up on animations because i'm a squirrelfly