attempt 3 done?
I think I've pushed as far as I can with this approach. I'm storing the name of the method the device sets an interval for as a string in a property of the device, then using object[key] notation to fire the method using the string. the problem here is that doing this i can't pass args into the action.

i'm not sure how to avoid writing myself into a corner with this. time to think i guess.